Refactored the Class to make it completely static, including events
fixed an issue in de-collide-console-spec where assertions were made before the console was un-hooked
enhanced the log output in the browser

added a test for three arg case

general tidy up to enhance encapsulation

Got rid of all of the noise during test. terminal mode is now completely clean

added a feature that ensures the Class can only have one instance.  Added a test for this

Added error objects in index and export them if NODE_ENV === 'test'
Added a module to suport this in the browser version as well